drug-interaction-checker/drug-interaction-checker/requirements.txt
streamlit>=1.18
requests>=2.28
langchain>=0.0.200    # your milage may vary; if not available, the app has fallbacks
faiss-cpu             # if pip install fails on some platforms, see README for fallback
python-dotenv
openai
tqdm
streamlit
langchain
langchain-openai
google-generativeai>=0.5.0
rich